relabel raf-db

+++ RafdbClass.py
@@ -157,7 +157,7 @@ class RafDB:
             if file.endswith(".jpg") or file.endswith(".png"):
                 if os.path.exists(os.path.join(self.anno_path_aug, file[:-4] + "_exp.npy")) \
                         and os.path.exists(os.path.join(self.anno_path_aug, file[:-4] + "_slnd.npy")):
-                    lbl = np.float64(np.load(os.path.join(self.anno_path_aug, file[:-4] + "_exp.npy")))
+                    lbl = np.int64(np.load(os.path.join(self.anno_path_aug, file[:-4] + "_exp.npy")))
                     lbl = lbl - 1
                     save(os.path.join(self.anno_path_aug, file[:-4] + "_exp.npy"), lbl)
